feature:cpu:model: use double instead of int

+++ lib/app/features/cpu/model/core_model.dart
@@ -3,7 +3,7 @@ import 'package:penguin_kernel_manager/app/features/cpu/model/core_state_model.d
 class CpuCore {
   final int coreNumber;
   final String currentFrequencyNode;
-  final List<int> availableFrequencies;
+  final List<double> availableFrequencies;
   final List<String> availableGovernor;
   final List<CoreStateModel> state;
   CpuCore(this.availableFrequencies, this.coreNumber, this.currentFrequencyNode,
